test: Inline scriptnum10 header
