+++ src/Makefile.test.include
@@ -30,7 +30,6 @@ BITCOIN_TEST_SUITE = \
 # test_bitcoin binary #
 BITCOIN_TESTS =\
   test/arith_uint256_tests.cpp \
-  test/scriptnum10.h \
   test/addrman_tests.cpp \
   test/amount_tests.cpp \
   test/allocator_tests.cpp \
